One of those Xcode "Update to recommended settings" nail-biters.
Xcode said that it found that codesigning was switched ON for Bkmxwork.framework, Debug and Release.  These certainly should not have been on.  I don't know how they got on.  Possibly because I just merged branch master into 2.5NewDoc, and as usual there caused some shenanigans in the Xcode project file.  So I said, "OK" and let it "update".  The result was that indeed these two codesign identifiers in the .pbxproj were changed to empty strings.  Also, it changed the LastUpgradeVersion in all of BkmkMgrs schemes, and all of Sparkle's schemes, from "0900" to "0910".  This is strange because I've been using Xcode 9.1 for several weeks at least.  Xcode, oh Xcode!!